Centralize profanity checking logic
- Move checkProfanityAPI() and checkEntryProfanity() to validation-utils
- Remove duplicate code from check-profanity.js and comprehensive-validation.js
- Improves maintainability and follows DRY principle
