+++ scripts/validation-utils.js
@@ -129,6 +129,49 @@ function sanitizeInput(input) {
   return input.replace(/[;&|`$(){}[\]<>]/g, '').trim();
 }
 
+async function checkProfanityAPI(text, retryCount = 0) {
+  const MAX_RETRIES = 3;
+  const TIMEOUT_MS = 5000;
+  const RETRY_DELAYS = [1000, 2000, 3000];
+
+  try {
+    const controller = new AbortController();
+    const timeoutId = setTimeout(() => controller.abort(), TIMEOUT_MS);
+
+    const response = await fetch(
+      `https://www.purgomalum.com/service/containsprofanity?text=${encodeURIComponent(text)}`,
+      { signal: controller.signal }
+    );
+
+    clearTimeout(timeoutId);
+    const result = await response.text();
+    return result.toLowerCase() === 'true';
+  } catch (error) {
+    console.log(`Profanity API check failed for "${text}" (attempt ${retryCount + 1}/${MAX_RETRIES}): ${error.message}`);
+
+    if (retryCount < MAX_RETRIES - 1) {
+      const delay = RETRY_DELAYS[retryCount];
+      console.log(`Retrying in ${delay}ms...`);
+      await new Promise(resolve => setTimeout(resolve, delay));
+      return checkProfanityAPI(text, retryCount + 1);
+    }
+
+    console.warn(`Profanity API unavailable after ${MAX_RETRIES} attempts. Flagging for manual review.`);
+    return true;
+  }
+}
+
+async function checkEntryProfanity(name, message) {
+  const profanityInName = name ? await checkProfanityAPI(name) : false;
+  const profanityInMessage = message ? await checkProfanityAPI(message) : false;
+
+  return {
+    hasProfanity: profanityInName || profanityInMessage,
+    profanityInName,
+    profanityInMessage
+  };
+}
+
 module.exports = {
   MAINTAINERS,
   ENTRY_REGEX,
@@ -138,5 +181,7 @@ module.exports = {
   parseSimpleFormat,
   validateFormat,
   checkExistingContributor,
-  sanitizeInput
+  sanitizeInput,
+  checkProfanityAPI,
+  checkEntryProfanity
 };
